+++ packages/virtual/core/src/index.ts
@@ -3,8 +3,6 @@
  * {@inheritDoc VirtualScroller}
  */
 
-import "@af-utils/scrollend-polyfill";
-
 export { default as VirtualScroller } from "models/VirtualScroller";
 export { VirtualScrollerEvent } from "constants/";
 export * from "./types";