+++ packages/website/components/examples/react-examples/hook/grid/meta.ts
@@ -1,0 +1,9 @@
+import { Metadata } from "next";
+
+const metadata = {
+    description:
+        "Virtualized dynamic-sized grids are supported, \
+just call useVirtual two times for rows and columns."
+} satisfies Metadata;
+
+export default metadata;